move ForcedTime and GrowattTimeCalendar
