+++ growatt_public_api/pydantic_models/min.py
@@ -1,10 +1,8 @@
 import datetime
-from typing import Union, Any, List, Optional, TypeAlias
+from typing import Union, Any, List
 
-from loguru import logger
 from pydantic import (
     ConfigDict,
-    BeforeValidator,
 )
 from pydantic.alias_generators import to_camel
 
@@ -13,10 +11,9 @@ from pydantic_models.api_model import (
     EmptyStrToNone,
     GrowattTime,
     ApiModel,
-    GrowattTimeGregorianChange,
+    GrowattTimeCalendar,
+    ForcedTime,
 )
-
-from typing import Annotated
 
 
 # #####################################################################################################################
@@ -50,22 +47,6 @@ def _min_tlx_settings_data_to_camel(snake: str) -> str:
         "win_mode_on_grid_discharge_stop_soc": "winModeOnGridDischargeStopSOC",
     }
     return override.get(snake, to_camel(snake=snake))
-
-
-def parse_forced_time(value: Optional[str] = None):
-    """support 0:0 for 00:00"""
-    if value and value.strip():
-        # noinspection PyBroadException
-        try:
-            return datetime.datetime.strptime(value, "%H:%M").time()
-        except:  # noqa: E722 do not use bare 'except'
-            logger.warning(f'invalid time "{value}" set to None')
-            return None
-    else:
-        return None
-
-
-ForcedTime: TypeAlias = Annotated[Union[datetime.time, None], BeforeValidator(parse_forced_time)]
 
 
 class MinTlxSettingsData(ApiModel):
@@ -531,45 +512,6 @@ class MinDetails(ApiResponse):
 
 # #####################################################################################################################
 # Min energy overview #################################################################################################
-
-
-def _growatt_time_calendar_timezone_to_camel(snake: str) -> str:
-    """
-    define own to_camel function to support weird API naming
-    """
-    override = {
-        "dst_savings": "DSTSavings",
-        "id": "ID",
-    }
-    return override.get(snake, to_camel(snake=snake))
-
-
-class GrowattTimeCalendarTimeZone(ApiModel):
-    model_config = ConfigDict(
-        from_attributes=True,
-        populate_by_name=True,
-        alias_generator=_growatt_time_calendar_timezone_to_camel,
-    )
-
-    dirty: Union[EmptyStrToNone, bool] = None  # e.g. false
-    display_name: Union[EmptyStrToNone, str] = None  # e.g. "China Standard Time"
-    dst_savings: Union[EmptyStrToNone, int] = None  # e.g. 0
-    id: Union[EmptyStrToNone, str] = None  # e.g. "Asia/Shanghai"
-    last_rule_instance: Union[EmptyStrToNone, str] = None  # e.g. null
-    raw_offset: Union[EmptyStrToNone, int] = None  # e.g. 28800000
-
-
-class GrowattTimeCalendar(ApiModel):
-    minimal_days_in_first_week: Union[EmptyStrToNone, int] = None  # e.g. 1
-    week_year: Union[EmptyStrToNone, int] = None  # e.g. 2018
-    time: Union[EmptyStrToNone, GrowattTime] = None
-    weeks_in_week_year: Union[EmptyStrToNone, int] = None  # e.g. 52
-    gregorian_change: Union[EmptyStrToNone, GrowattTimeGregorianChange] = None
-    time_zone: Union[EmptyStrToNone, GrowattTimeCalendarTimeZone] = None
-    time_in_millis: Union[EmptyStrToNone, int] = None  # e.g. 1544670232000
-    lenient: Union[EmptyStrToNone, bool] = None  # e.g. true
-    first_day_of_week: Union[EmptyStrToNone, int] = None  # e.g. 1
-    week_date_supported: Union[EmptyStrToNone, bool] = None
 
 
 def _min_energy_overview_data_to_camel(snake: str) -> str:
